Updated debug ArrayView visualizers to support new ArrayView functionality.
